requirements.txt
# Django Core
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers

# Database
psycopg2-binary

# Celery & Redis
celery
redis
django-celery-beat

# Email
django-templated-mail

# Utils
Pillow
python-decouple
drf-spectacular

# Development
gunicorn
# .env file support
python-dotenv

# HTTP Requests
requests
# Static Files Handling
whitenoise
